feat: Phase 7g — unified home page + bug fixes (#344, #345, #348)
Merge QuickHomeScreen into HomePage so both modes share one home page.
The Quick/Full distinction now only matters inside trip views.

- Merge scan CTA, QuickCreateSheet, motion animations into HomePage
- Simplify ConditionalHomePage: only redirect on mobile + active trip
- Remove /quick home route, add redirect for backward compat
- Fix "Full view" pill navigating to /expenses instead of /dashboard (#345)
- Fix QuickLayout back arrow pointing to /quick instead of /
- Hide scan + mode toggle from Layout header on home page
- Mode toggle: only navigate when inside a trip
- Delete QuickHomeScreen.tsx (all features merged into HomePage)
- Update unit + e2e tests
