--- src/pages/QuickHomeScreen.tsx
@@ -1,233 +1,0 @@
-import { useState, useCallback, useMemo } from 'react'
-import { useNavigate } from 'react-router-dom'
-import { useAuth } from '@/contexts/AuthContext'
-import { useTripContext } from '@/contexts/TripContext'
-import { useMyTripBalances } from '@/hooks/useMyTripBalances'
-import { getHiddenTripCodes, showTrip } from '@/lib/mutedTripsStorage'
-import { getActiveTripId } from '@/lib/activeTripDetection'
-import { GroupActions } from '@/components/quick/GroupActions'
-import { QuickCreateSheet } from '@/components/quick/QuickCreateSheet'
-import { QuickScanContextSheet } from '@/components/quick/QuickScanContextSheet'
-import { QuickScanCreateFlow } from '@/components/quick/QuickScanCreateFlow'
-import { TripCard } from '@/components/TripCard'
-import { PageLoadingState } from '@/components/PageLoadingState'
-import { Card, CardContent } from '@/components/ui/card'
-import { ChevronRight, Plus, Eye, ChevronDown, ExternalLink, ScanLine } from 'lucide-react'
-import { Button } from '@/components/ui/button'
-import { motion } from 'framer-motion'
-
-export function QuickHomeScreen() {
-  const navigate = useNavigate()
-  const { user, userProfile } = useAuth()
-  const { loading: tripsLoading } = useTripContext()
-  const { tripBalances, loading: balancesLoading } = useMyTripBalances()
-  const loading = balancesLoading || (!!user && tripsLoading)
-  const [hiddenCodes, setHiddenCodes] = useState(() => new Set(getHiddenTripCodes()))
-  const [showHidden, setShowHidden] = useState(false)
-  const [createOpen, setCreateOpen] = useState(false)
-  const [scanContextOpen, setScanContextOpen] = useState(false)
-  const [scanCreateOpen, setScanCreateOpen] = useState(false)
-
-  const firstName = userProfile?.display_name?.split(' ')[0] || 'there'
-
-  const visibleTrips = tripBalances.filter(tb => !hiddenCodes.has(tb.trip.trip_code))
-  const hiddenTrips = tripBalances.filter(tb => hiddenCodes.has(tb.trip.trip_code))
-
-  const activeTripId = useMemo(
-    () => getActiveTripId(visibleTrips.map(tb => tb.trip)),
-    [visibleTrips]
-  )
-
-  const handleHidden = useCallback((tripCode: string) => {
-    setHiddenCodes(prev => new Set([...prev, tripCode]))
-  }, [])
-
-  const handleShow = useCallback((tripCode: string) => {
-    showTrip(tripCode)
-    setHiddenCodes(prev => {
-      const next = new Set(prev)
-      next.delete(tripCode)
-      return next
-    })
-  }, [])
-
-  const handleLeft = useCallback((tripCode: string) => {
-    setHiddenCodes(prev => new Set([...prev, tripCode]))
-  }, [])
-
-  const handleScanTap = () => {
-    if (loading) return
-    if (visibleTrips.length === 0) {
-      // No groups — go straight to scan+create flow
-      setScanCreateOpen(true)
-    } else {
-      // 1+ groups — always show picker so user can choose or create new
-      setScanContextOpen(true)
-    }
-  }
-
-  return (
-    <div className="min-h-screen bg-background">
-      <div className="max-w-lg mx-auto px-4 py-8">
-        {/* User greeting */}
-        <div className="flex items-center gap-3 mb-6">
-          {userProfile?.avatar_url ? (
-            <img
-              src={userProfile.avatar_url}
-              alt={userProfile.display_name}
-              className="w-12 h-12 rounded-full"
-              referrerPolicy="no-referrer"
-            />
-          ) : (
-            <div className="w-12 h-12 rounded-full bg-primary/10 flex items-center justify-center text-lg font-semibold text-primary">
-              {firstName[0]?.toUpperCase()}
-            </div>
-          )}
-          <div>
-            <h1 className="text-2xl font-bold text-foreground">
-              Hi, {firstName}
-            </h1>
-            <p className="text-sm text-muted-foreground">Your groups</p>
-          </div>
-        </div>
-
-        {/* Scan CTA */}
-        <button
-          onClick={handleScanTap}
-          disabled={loading}
-          className="w-full flex items-center justify-between px-5 py-4 rounded-2xl bg-primary text-primary-foreground hover:bg-primary/90 active:scale-[0.98] transition-all mb-6 disabled:opacity-50"
-        >
-          <div className="flex items-center gap-3">
-            <div className="w-9 h-9 rounded-full bg-primary-foreground/20 flex items-center justify-center">
-              <ScanLine size={20} />
-            </div>
-            <div className="text-left">
-              <p className="font-semibold text-sm leading-tight">Scan a Receipt</p>
-              <p className="text-xs text-primary-foreground/70 mt-0.5">
-                {visibleTrips.length === 0 ? 'Creates a new group automatically' : 'Add expenses from a photo'}
-              </p>
-            </div>
-          </div>
-          <ChevronRight size={18} className="opacity-70" />
-        </button>
-
-        {/* Groups list */}
-        {loading ? (
-          <PageLoadingState />
-        ) : visibleTrips.length === 0 ? (
-          <Card>
-            <CardContent className="pt-6">
-              <div className="text-center py-8">
-                <p className="text-muted-foreground mb-4">
-                  No groups yet. Create a trip or event, or join one via a shared link.
-                </p>
-                <Button onClick={() => setCreateOpen(true)} className="gap-2">
-                  <Plus size={18} />
-                  New Trip or Event
-                </Button>
-              </div>
-            </CardContent>
-          </Card>
-        ) : (
-          <div className="space-y-3">
-            {visibleTrips.map(({ trip, myBalance }, i) => (
-              <motion.div
-                key={trip.id}
-                initial={{ opacity: 0, y: 10 }}
-                animate={{ opacity: 1, y: 0 }}
-                transition={{ delay: i * 0.05 }}
-              >
-                <TripCard
-                  trip={trip}
-                  balance={myBalance}
-                  isActive={trip.id === activeTripId}
-                  onClick={() => navigate(`/t/${trip.trip_code}/quick`)}
-                  actions={
-                    <GroupActions
-                      tripCode={trip.trip_code}
-                      tripName={trip.name}
-                      onHidden={() => handleHidden(trip.trip_code)}
-                      onLeft={() => handleLeft(trip.trip_code)}
-                    />
-                  }
-                />
-              </motion.div>
-            ))}
-          </div>
-        )}
-
-        {/* Create trip button */}
-        {visibleTrips.length > 0 && (
-          <div className="mt-6 text-center">
-            <Button variant="outline" onClick={() => setCreateOpen(true)} className="gap-2">
-              <Plus size={16} />
-              New Trip or Event
-            </Button>
-          </div>
-        )}
-
-        {/* What's New link */}
-        <div className="mt-8 text-center">
-          <a
-            href="https://github.com/kristjanelias-xtian/trip-splitter-pro/blob/main/RELEASE_NOTES.md"
-            target="_blank"
-            rel="noopener noreferrer"
-            className="inline-flex items-center gap-1.5 text-xs text-muted-foreground hover:text-foreground transition-colors"
-          >
-            What's New · v0.9.4
-            <ExternalLink size={12} />
-          </a>
-        </div>
-
-        {/* Hidden trips section */}
-        {hiddenTrips.length > 0 && (
-          <div className="mt-8">
-            <button
-              onClick={() => setShowHidden(prev => !prev)}
-              className="flex items-center gap-2 text-sm text-muted-foreground hover:text-foreground transition-colors w-full"
-            >
-              <ChevronDown
-                size={16}
-                className={`transition-transform ${showHidden ? 'rotate-180' : ''}`}
-              />
-              {hiddenTrips.length} hidden {hiddenTrips.length === 1 ? 'group' : 'groups'}
-            </button>
-
-            {showHidden && (
-              <div className="mt-3 space-y-2">
-                {hiddenTrips.map(({ trip }) => (
-                  <div
-                    key={trip.id}
-                    className="flex items-center justify-between px-4 py-3 rounded-lg border border-border bg-muted/30"
-                  >
-                    <span className="text-sm text-muted-foreground truncate">
-                      {trip.name}
-                    </span>
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      className="gap-1.5 text-muted-foreground flex-shrink-0"
-                      onClick={() => handleShow(trip.trip_code)}
-                    >
-                      <Eye size={14} />
-                      Show
-                    </Button>
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-        )}
-      </div>
-
-      <QuickCreateSheet open={createOpen} onOpenChange={setCreateOpen} />
-      <QuickScanContextSheet
-        open={scanContextOpen}
-        onOpenChange={setScanContextOpen}
-        trips={visibleTrips.map(tb => tb.trip)}
-        onNewGroup={() => setScanCreateOpen(true)}
-      />
-      <QuickScanCreateFlow open={scanCreateOpen} onOpenChange={setScanCreateOpen} />
-    </div>
-  )
-}
